make `%TypedArray%.prototype.sort` stricter
