+++ tests/tests/es.array.sort.js
@@ -80,6 +80,8 @@ QUnit.test('Array#sort', assert => {
 
   array.sort((a, b) => (a / 4 | 0) - (b / 4 | 0));
 
+  assert.ok(1 / [0, -0].sort()[0] > 0, '-0');
+
   assert.same(String(array), String(expected), 'stable #1');
 
   let result = '';
